--- a/client/src/data/repository/underlying_repo.rs
+++ b/client/src/data/repository/underlying_repo.rs
@@ -22,7 +22,9 @@ impl UnderlyingRepo for UnderlyingRepoImpl {
     async fn forward(&self, request: Vec<u8>, host: String) -> Result<Vec<u8>, String> {
         //info!("Forwarding request: {} to host: {}", String::from_utf8(request.clone()).unwrap(), host.clone());
         let stream = TcpStream::connect(host.as_str()).await.unwrap();
-        let mut stream = TcpStreamTLS::from_tcp(stream);
+        let (read_stream, write_stream) = tokio::io::split(stream);
+        let mut stream = TcpStreamTLS::from_tcp(read_stream, write_stream);
+        
         // forward request
         stream.write_all(&request).await.unwrap();