--- a/client/src/handler/main_handler.rs
+++ b/client/src/handler/main_handler.rs
@@ -1,5 +1,3 @@
-// TODO: implement this
-
 use std::{sync::Arc, time::Duration};
 
 use http::StatusCode;
@@ -7,10 +5,12 @@ use http::StatusCode;
 use common::{
     convert::{from_json_slice, to_json_vec}, 
     data::dto::{public_request::PublicRequest, public_response::PublicResponse, tunnel_client::TunnelClient}, 
-    net::{ack_health_check_packet, http_json_response_as_bytes, read_bytes_from_mutexed_socket, read_string_from_socket, HttpResponse, TcpStreamTLS}, 
+    net::{
+        http_json_response_as_bytes, prepare_packet, read_bytes_from_mutexed_socket, read_string_from_socket, separate_packets, HttpResponse, TcpStreamTLS, HEALTH_CHECK_PACKET_ACK
+    }, 
     security::sign_value
 };
-use tokio::{net::TcpStream, sync::Mutex, time::{sleep, Instant}};
+use tokio::{net::TcpStream, sync::{Mutex, mpsc, mpsc::{Sender, Receiver}}, time::{sleep, Instant}};
 use log::{error, info};
 use tokio_native_tls::{native_tls, TlsConnector};
 
@@ -40,7 +40,7 @@ pub async fn register_handler(underlying_host: String, service: UnderlyingServic
                 continue;
             }
         };
-        let mut stream = if use_tls {
+        let (mut read_stream, mut write_stream) = if use_tls {
             let cert = get_ca_certificate().unwrap();
             let connector = native_tls::TlsConnector::builder()
                 .add_root_certificate(cert)
@@ -48,30 +48,26 @@ pub async fn register_handler(underlying_host: String, service: UnderlyingServic
                 .unwrap();
             let connector = TlsConnector::from(connector);
             let tls_stream = connector.connect(server_host.as_str(), tcp_stream).await.unwrap();
+            let (read_stream, write_stream) = tokio::io::split(tls_stream);
             info!("TLS Bound -> address: {}", server_address.clone());
-            TcpStreamTLS {
-                tcp: None,
-                tls: Some(tls_stream)
-            }
+            (TcpStreamTLS::from_tcp_tls_read(read_stream), TcpStreamTLS::from_tcp_tls_write(write_stream))
         } else { 
-            TcpStreamTLS {
-                tcp: Some(tcp_stream),
-                tls: None
-            }
+            let (read_stream, write_stream) = tokio::io::split(tcp_stream);
+            (TcpStreamTLS::from_tcp_read(read_stream), TcpStreamTLS::from_tcp_write(write_stream))
          };
         // send connection request to server service
         let tunnel_client = get_tunnel_client();
         let bytes_req = to_json_vec(&tunnel_client);
 
         info!("Connecting to server service...");
-        if let Err(e) = stream.write_all(&bytes_req).await {
+        if let Err(e) = write_stream.write_all(&bytes_req).await {
             error!("Error connecting to server service: {}", e);
             continue;
         }
 
         // check if the server handshake was successful
         let mut ok: String = Default::default();
-        if let Err(e) = read_string_from_socket(&mut stream, &mut ok).await {
+        if let Err(e) = read_string_from_socket(&mut read_stream, &mut ok).await {
             error!("Error connecting to server service: {}", e);
             continue;
         }
@@ -82,9 +78,31 @@ pub async fn register_handler(underlying_host: String, service: UnderlyingServic
 
         info!("Connected to server service.");
 
-        let stream_mutex = Arc::new(Mutex::new(stream));
-        // start tunnel handler
-        tunnel_handler(stream_mutex, underlying_host.clone(), service.clone()).await;
+        
+        // create channel for request queue
+        let (tx, rx) = mpsc::channel::<PublicResponse>(5);
+
+        // convert to mutex
+        let tx_mutex = Arc::new(Mutex::new(tx));
+        let rx_mutex = Arc::new(Mutex::new(rx));
+        let read_stream_mutex = Arc::new(Mutex::new(read_stream));
+        let write_stream_mutex = Arc::new(Mutex::new(write_stream));
+        
+        // TODO: add break flag if one of the handlers stopped (?)
+
+        // spawn handlers
+        let cloned_underlying_host = underlying_host.clone();
+        let cloned_service = service.clone();
+        let receiver_handler = tokio::spawn(async move {
+            tunnel_receiver_handler(read_stream_mutex, tx_mutex, cloned_underlying_host, cloned_service).await;
+        });
+        let sender_handler = tokio::spawn(async move {
+            tunnel_sender_handler(write_stream_mutex, rx_mutex).await;
+        });
+
+        // wait until released
+        receiver_handler.await.unwrap_or_default();
+        sender_handler.await.unwrap_or_default();
     }
 }
 
@@ -97,8 +115,8 @@ fn get_tunnel_client() -> TunnelClient {
     TunnelClient::new(client_id, signature)
 }
 
-pub async fn tunnel_handler(stream: Arc<Mutex<TcpStreamTLS>>, underlying_host: String, service: UnderlyingService) {
-    info!("Tunnel handler started.");
+pub async fn tunnel_receiver_handler(stream: Arc<Mutex<TcpStreamTLS>>, tx: Arc<Mutex<Sender<PublicResponse>>>, underlying_host: String, service: UnderlyingService) {
+    info!("Tunnel receiver handler started.");
     loop {
         // get incoming request server service to forward
         let mut request = Vec::new();
@@ -112,45 +130,68 @@ pub async fn tunnel_handler(stream: Arc<Mutex<TcpStreamTLS>>, underlying_host: S
             continue;
         }
 
-        if ack_health_check_packet(stream.clone(), request.clone()).await {
-            // skip if the packet
-            continue;
-        }
-
-        let public_request: PublicRequest = from_json_slice(&request).unwrap();
-        let start_request = Instant::now();
-        info!("Incoming request: {} received, forwarding to underlying service...", public_request.id);
-        // forward response to underlying service
-        let res = service.foward_request(public_request.data, underlying_host.clone()).await;
-        if res.is_err() {
-            // response error to server
-            let msg = String::from("Request cannot be processed");
-            let response = match http_json_response_as_bytes(
-                HttpResponse::new(false, msg), StatusCode::from_u16(400).unwrap()) {
-                    Ok(value) => value,
-                    Err(err) => {
-                        info!("error: {}", err);
-                        continue;
-                    } 
+        let packets = separate_packets(request);
+        for packet in packets {
+            let public_request: PublicRequest = from_json_slice(&packet).unwrap(); // assuming correct format
+            let start_request = Instant::now();
+            info!("Incoming request: {} received, forwarding to underlying service...", public_request.id);
+            
+            // dispatch request to underlying service
+            let cloned_underlying_host = underlying_host.clone();
+            let cloned_service = service.clone();
+            let cloned_tx = tx.clone();
+            tokio::spawn(async move {
+                let public_response = match cloned_service.foward_request(public_request.data, cloned_underlying_host).await {
+                    Ok(res) => {
+                        PublicResponse::new(public_request.id.clone(), res.clone())
+                    },
+                    Err(_) => {
+                        let msg = String::from("Request cannot be processed");
+                        let res = http_json_response_as_bytes(
+                            HttpResponse::new(false, msg), StatusCode::from_u16(400).unwrap()).unwrap();
+                        PublicResponse::new(public_request.id.clone(), res.clone())
+                    }
                 };
-
-            stream.lock().await.write_all(&response).await.unwrap();
-
-            continue;
-        }
-
-        info!("Response for request {} received in {} seconds.", public_request.id,start_request.elapsed().as_secs());
         
-        let res = res.unwrap();
-        let public_response = PublicResponse::new(public_request.id.clone(), res.clone());
-
-        // foward response from underlying service to server service
-        let bytes_res = to_json_vec(&public_response);
-        if let Err(e) = stream.lock().await.write_all(&bytes_res).await {
-            error!("{}", e);
-            break;
+                if let Ok(_) = cloned_tx.lock().await.send(public_response).await {
+                    info!("Response for request {} received in {} seconds and was enqueued to foward back", public_request.id, start_request.elapsed().as_secs());
+                }
+            });
         }
-
-        info!("Incoming request: {} processed.", public_request.id);
     }
+
+    error!("Connection closed.");
+}
+
+pub async fn tunnel_sender_handler(stream: Arc<Mutex<TcpStreamTLS>>, rx: Arc<Mutex<Receiver<PublicResponse>>>) {
+    info!("Tunnel sender handler started.");
+    let mut skip = 0;
+    loop {
+        // get ready public responses from the queue
+        if let Some(public_response) = rx.lock().await.recv().await {
+            info!("Response for request: {} is available.", public_response.request_id);
+            // foward response from underlying service to server service
+            let bytes_res = prepare_packet(to_json_vec(&public_response));
+            if let Err(e) = stream.lock().await.write_all(&bytes_res).await {
+                error!("{}", e);
+                break;
+            }
+
+            info!("Request: {} processed.", public_response.request_id);
+        } else {
+            skip += 1;
+            // every 20k skips send health check
+            if skip == 20000 {
+                let hc = prepare_packet(Vec::from(String::from(HEALTH_CHECK_PACKET_ACK).as_bytes()));
+                if let Err(_) = stream.lock().await.write_all(&hc).await {
+                    break;
+                }
+                // sleep for 0.5 seconds
+                sleep(Duration::from_millis(100)).await;
+                skip = 0;
+            }
+        }
+    }
+
+    error!("Connection closed.");
 }