--- a/server/src/data/repository/request_repo.rs
+++ b/server/src/data/repository/request_repo.rs
@@ -7,11 +7,12 @@ const REDIS_KEY_PENDING_PUBLIC_REQUEST: &str = "pending_public_requests";
 
 #[async_trait]
 pub trait RequestRepo {
-    async fn push_back(&self, request: PublicRequest) -> Result<(), String>;
+    async fn push_back(&self, client_id: String, request: PublicRequest) -> Result<(), String>;
     async fn pop_front(&self, client_id: String) -> Result<PublicRequest, String>;
     async fn queue_len(&self, client_id: String) -> Result<u16, String>;
     async fn ack_pending(&self, client_id: String, request_id: String) -> Result<(), String>;
     async fn ack_done(&self, client_id: String, request_id: String) -> Result<(), String>;
+    async fn is_pending(&self, client_id: String, request_id: String) -> bool;
 }
 
 pub struct RequestRepoImpl {
@@ -26,9 +27,9 @@ impl RequestRepoImpl {
 
 #[async_trait]
 impl RequestRepo for RequestRepoImpl {
-    async fn push_back(&self, request: PublicRequest) -> Result<(), String> {
+    async fn push_back(&self, client_id: String, request: PublicRequest) -> Result<(), String> {
         let data = to_json_vec(&request);
-        let key = format!("{}_{}", REDIS_KEY_PUBLIC_REQUEST, request.client_id);
+        let key = format!("{}_{}", REDIS_KEY_PUBLIC_REQUEST, client_id);
         self.connection.clone().lpush(key, &data).await
             .map_err(|e| format!("Error pushing request {}: {}", request.id, e))?;
         Ok(())
@@ -67,5 +68,11 @@ impl RequestRepo for RequestRepoImpl {
         self.connection.clone().hdel(key, request_id.clone()).await
             .map_err(|e| format!("Error unsetting pending request {}: {}", request_id, e))?;
         Ok(())
+    }
+
+    async fn is_pending(&self, client_id: String, request_id: String) -> bool {
+        let key = format!("{}_{}", REDIS_KEY_PENDING_PUBLIC_REQUEST, client_id);
+        let data: Vec<u8> = self.connection.clone().hget(key, request_id.clone()).await.unwrap_or_default();
+        return data.len() > 0
     }
 }