--- a/server/src/handler/tunnel_handler.rs
+++ b/server/src/handler/tunnel_handler.rs
@@ -149,6 +149,7 @@ async fn tunnel_sender_handler(stream: Arc<Mutex<TcpStreamTLS>>, public_service:
     
     // disconnection
     client_service.lock().await.disconnect_client(client_id.clone()).await.unwrap();
+    info!("Tunnel sender handler stopped.");
     info!("Client Disconnected. client_id: {}", client_id);
 }
 
@@ -189,5 +190,6 @@ async fn tunnel_receiver_handler(stream: Arc<Mutex<TcpStreamTLS>>, public_servic
     
     // disconnection
     client_service.lock().await.disconnect_client(client_id.clone()).await.unwrap();
+    info!("Tunnel receiver handler stopped.");
     info!("Client Disconnected. client_id: {}", client_id);
 }