**Changes:**
- Improved performance for handling multiple requests

**Benchmark**
Tested with:
50 batches with 40 requests sent and a 400ms delay per batch.

Before:
```
Run 1:
Elapsed time: 66.2148 seconds
request_cnt: 2000
success_cnt:  633

Run 2:
Elapsed time: 66.8324 seconds
request_cnt: 2000
success_cnt:  600

Run 3:
Elapsed time: 66.9765 seconds
request_cnt: 2000
success_cnt:  516
```

After:
```
Run 1:
Elapsed time: 62.5620 seconds
request_cnt: 2000
success_cnt:  1578

Run 2:
Elapsed time: 31.4077 seconds
request_cnt: 2000
success_cnt:  2000

Run 3:
Elapsed time: 45.6027 seconds
request_cnt: 2000
success_cnt:  2000
```

*The time out for each request is 30 seconds, so the unsuccessful ones expire.  It prints the error response, which also affects the overall elapsed time.